refactor: 레싸바 장례식ㅠ
