+++ src/shared/component/SideNavBar/PageIndicatorStick/PageIndicatorStick.style.ts
